ART: dex2oat and instruction_set_features: utilize modified runtime capabilities set
For some time, it bothered me that the device is actively trying to target different to known cpudefines set.
When inspecting clang's defines system, it occured to me that when we're supplying mcpu and march, the gaps aren't filled to their fullest.
e.g. march=armv8.2-a mcpu=cortex-a55 would skip over the fact that CPU supports fp16, lse, dotprod etc.
     ^ AOSP default(!)

Let's instead depend on hwcaps intersected with CPP defines - that way, we can at least presume the target system, with actually supplied hwcaps being the preference in order to accomplish the result of optimizing for the actual hardware at use.
ARM64 only, but I don't think the problem is as severe on anything else.

Change-Id: I1f631c4e3ae1b42f7e4fc0aedf5d642186c734db
Signed-off-by: XeonDead <xeondead@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>